Smarter detection of whether we need to regenerate our outputs.
We now check the mtime of all the input images and regenerate resized outputs if
needed. In addition, we check the mtime on the input HAML template (and all its
dependencies, identified runtie), and regenerate any albums that have not been
updated since the templates were modified.

Add comments to Album and to the main workflow.